WIP: Basic Selection and Schematic Systems + Expanded Shape Configuration + Basic Menu Bar + More Fixes and Updates

+++ src/main/java/me/andre111/voxedit/tool/shape/Cube.java
@@ -19,7 +19,7 @@ import net.minecraft.util.math.Direction;
 
 public class Cube extends Shape {
 	@Override
-	public boolean contains(int x, int y, int z, Direction direction, int sizeX, int sizeY, int sizeZ) {
+	public boolean contains(int x, int y, int z, Direction direction, double sizeX, double sizeY, double sizeZ) {
 		return true;
 	}
 }
